requirements.txt
modelscope==1.3.0
torch==1.12.1
tqdm==4.64.1
sacrebleu==2.2.1
transformers==4.10.0
rouge==1.0.1
faiss-cpu
scikit-learn
sentencepiece
ujson